requirements.txt
streamlit
pandas
requests
akamai-edgegrid
plotly
openpyxl